requirements.txt
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.104.0
requests==2.31.0
python-telegram-bot==20.7
pytz==2024.1